ai-assistants/requirements.txt
openai>=1.3.0
slack-bolt>=1.18.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0 